Fixed some calibration bugs
Detects calibration pattern, but still not 100% ready for single cam
calibration though.

+++ include/calibrator/calibration.hpp
@@ -189,7 +189,7 @@ bool findPatchCorners(const cv::Mat& image, cv::Size patternSize, cv::Mat& homog
 bool findMaskCorners(const cv::Mat& image, cv::Size patternSize, vector<cv::Point2f>& corners, int detector = 0, int mserDelta = 8, double max_var = 0.25, double min_div = 0.2, double area_threshold = 1.01);
 
 /// \brief 		Core pattern-finding function
-bool findPatternCorners(const cv::Mat& image, cv::Size patternSize, vector<cv::Point2f>& corners, int mode, int detector = 0, int mserDelta = 8, float max_var = 0.25, float min_div = 0.2, double area_threshold = 1.01);
+bool findPatternCorners(const cv::Mat& image, cv::Size patternSize, vector<cv::Point2f>& corners, int mode, int detector = 0, int mserDelta = 8, double max_var = 0.25, double min_div = 0.2, double area_threshold = 1.01);
 
 /// \brief 		Find all patches (MSERS - using default settings) in an image
 void findAllPatches(const cv::Mat& image, cv::Size patternSize, vector<vector<cv::Point> >& msers, int mserDelta = 8, float max_var = 0.25, float min_div = 0.2, double area_threshold = 1.01);
